+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Change commands that run the developer tools to require the code to be loaded first
 * Inform the user when commands that spawn processes complete
 * Add experimental code lens support for test results
+* Add "Logtalk: Toggle Code Lens" command
 * Add "Logtalk: Generate Documentation (workspace)" command
 * Add "Logtalk: Generate Diagrams (workspace)" command
 * Add "Logtalk: Scan Dead Code (workspace)" command